Sign in done, now returns a jwt token with expiry = 1 min.
